Added the digital part of the project. (in progress)
I added the (so far) progress of the digital part. The CTRL_ex_time module is done and the Timer_counter is almost done. The testbench is done and everything is connected to it. It simulates the inputs but not yet the outputs. This is due to the State Machine not being done.